Move files measures/base -> measure/base
