+++ include/fullscore/models/measure/base.h
